# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Authoritative room and match APIs for Penalty Grid.
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type CreateRoomInputMode = typeof CreateRoomInputMode[keyof typeof CreateRoomInputMode];


export const CreateRoomInputMode = {
  classic: 'classic',
  party: 'party',
} as const;

export type CreateRoomInputMatchLength = typeof CreateRoomInputMatchLength[keyof typeof CreateRoomInputMatchLength];


export const CreateRoomInputMatchLength = {
  NUMBER_3: 3,
  NUMBER_5: 5,
  NUMBER_7: 7,
} as const;

export type CreateRoomInputTurnTimerSeconds = typeof CreateRoomInputTurnTimerSeconds[keyof typeof CreateRoomInputTurnTimerSeconds];


export const CreateRoomInputTurnTimerSeconds = {
  NUMBER_8: 8,
  NUMBER_12: 12,
  NUMBER_20: 20,
  NUMBER_30: 30,
} as const;

export type CreateRoomInputMaxPlayers = typeof CreateRoomInputMaxPlayers[keyof typeof CreateRoomInputMaxPlayers];


export const CreateRoomInputMaxPlayers = {
  NUMBER_2: 2,
  NUMBER_4: 4,
  NUMBER_6: 6,
  NUMBER_8: 8,
  NUMBER_10: 10,
} as const;

export type CreateRoomInputDifficulty = typeof CreateRoomInputDifficulty[keyof typeof CreateRoomInputDifficulty];


export const CreateRoomInputDifficulty = {
  casual: 'casual',
  competitive: 'competitive',
  hardcore: 'hardcore',
} as const;

export interface CreateRoomInput {
  /**
     * @minLength 2
     * @maxLength 18
     */
  displayName: string;
  /** @maxLength 2 */
  avatar?: string;
  mode?: CreateRoomInputMode;
  matchLength?: CreateRoomInputMatchLength;
  turnTimerSeconds?: CreateRoomInputTurnTimerSeconds;
  maxPlayers?: CreateRoomInputMaxPlayers;
  difficulty?: CreateRoomInputDifficulty;
}

export interface JoinRoomInput {
  /**
     * @minLength 2
     * @maxLength 18
     */
  displayName: string;
  /** @maxLength 2 */
  avatar?: string;
}

export interface QuickPlayInput {
  /**
     * @minLength 2
     * @maxLength 18
     */
  displayName: string;
  /** @maxLength 2 */
  avatar?: string;
}

export type PlayerSnapshotTeam = typeof PlayerSnapshotTeam[keyof typeof PlayerSnapshotTeam];


export const PlayerSnapshotTeam = {
  home: 'home',
  away: 'away',
  neutral: 'neutral',
} as const;

export interface PlayerSnapshot {
  id: string;
  displayName: string;
  avatar: string;
  team: PlayerSnapshotTeam;
  ready: boolean;
  connected: boolean;
  score: number;
}

export type RoomSnapshotStatus = typeof RoomSnapshotStatus[keyof typeof RoomSnapshotStatus];


export const RoomSnapshotStatus = {
  lobby: 'lobby',
  playing: 'playing',
  finished: 'finished',
} as const;

export type RoomSnapshotMode = typeof RoomSnapshotMode[keyof typeof RoomSnapshotMode];


export const RoomSnapshotMode = {
  classic: 'classic',
  party: 'party',
} as const;

export type RoomSnapshotDifficulty = typeof RoomSnapshotDifficulty[keyof typeof RoomSnapshotDifficulty];


export const RoomSnapshotDifficulty = {
  casual: 'casual',
  competitive: 'competitive',
  hardcore: 'hardcore',
} as const;

export interface RoomSnapshot {
  code: string;
  hostId: string;
  status: RoomSnapshotStatus;
  mode: RoomSnapshotMode;
  matchLength: number;
  turnTimerSeconds: number;
  maxPlayers: number;
  difficulty: RoomSnapshotDifficulty;
  players: PlayerSnapshot[];
}

export interface PlayerSuggestion {
  id: string;
  name: string;
  imageUrl: string;
  aliases: string[];
}

export type PlayerSearchResponseFreshness = typeof PlayerSearchResponseFreshness[keyof typeof PlayerSearchResponseFreshness];


export const PlayerSearchResponseFreshness = {
  live: 'live',
  'verified-snapshot': 'verified-snapshot',
  stale: 'stale',
  unavailable: 'unavailable',
} as const;

export interface PlayerSearchResponse {
  players: PlayerSuggestion[];
  updatedAt: string;
  source: string;
  freshness: PlayerSearchResponseFreshness;
}

export interface RoomSession {
  code: string;
  playerId: string;
  sessionToken: string;
  room: RoomSnapshot;
}

export type SearchPlayersParams = {
/**
 * @maxLength 40
 */
search?: string;
limit?: SearchPlayersLimit;
};

export type SearchPlayersLimit = typeof SearchPlayersLimit[keyof typeof SearchPlayersLimit];


export const SearchPlayersLimit = {
  NUMBER_5: 5,
  NUMBER_8: 8,
  NUMBER_10: 10,
} as const;